Support python3
This work makes the necessary updates to the code to support both
Python2 and Python3.  To support this effort, updates to testing using
tox was implemented as an easy way to test compatibility in both
version.  This work is relatively clean, though in the test_result case,
I couldn't end up working out the correct incantation of .encode(),
BytesIO and StringIO to make each test pass, so its a hardcoded switch
to get the tests to pass.  I don't expect this to adversely effect the
usefulness of this code.