Fix argparse transition for positional arguments
The move from optparse to argparse seemed to have ignored the fact
that argparse will raise an error on unconsumed arguments, which
includes positional arguments.  The option parser should have
been sent sys.argv[1:] long ago, but I guess that was missed as
optparse would not complaing about the positional argument of the
script itself.  Just remove passing an args around and let the
option parser do the default of consuming sys.argv[1:]

Also did no one ever use positional arguments with the optparse
version?  Did no one test the argparse version?  Concerned we
are missing something here and there's not a bug.

Also the original github repo appears to be deleted so that
means no access to old github issues so the issue numbers
referenced in the argparse commit are of no help.